+++ Assets/Scripts/Mobs/IDamageable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void CalculateDamage(ref int damage);
-    public void ApplyDamage(int damage);
+    public void CalculateDamage(ref float damage);
+    public void ApplyDamage(float damage);
     public void CheckState();
 }
